requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch==2.4.0+cpu
torchvision==0.15.0+cpu
torchaudio==2.4.0+cpu
numpy
pandas
Flask
scikit-learn
requests
Pillow
gunicorn==20.0.4


# numpy
# pandas
# Flask
# scikit-learn
# https://download.pytorch.org/whl/cpu/torch-1.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
# https://download.pytorch.org/whl/cpu/torchvision-0.8.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
# requests
# Pillow
# gunicorn == 20.0.4